SignalR: refactoring to use the new package
