+++ azurerm/internal/services/signalr/client/client.go
@@ -1,4 +1,4 @@
-package signalr
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/signalr/mgmt/2018-10-01/signalr"
@@ -9,7 +9,7 @@ type Client struct {
 	Client *signalr.Client
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	client := signalr.NewClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&client.Client, o.ResourceManagerAuthorizer)
